simple global leaderboard added
for testing, not secure; better to use e.g. Google Play Services
