+++ Assets/_Scripts/UI/MenuManager.cs
@@ -22,6 +22,7 @@ namespace BlackHole
         [SerializeField] private Menu AchievementsMenuPrefab;
         [SerializeField] private Menu SettingsMenuPrefab;
         [SerializeField] private Menu HighscoresMenuPrefab;
+        [SerializeField] private Menu LeaderboardMenuPrefab;
 
         [SerializeField] private Transform _menuParent;
